+++ db/migrate/20190915021622_add_dm_reward_choice_to_log_entries.rb
@@ -1,0 +1,5 @@
+class AddDmRewardChoiceToLogEntries < ActiveRecord::Migration[5.2]
+  def change
+    add_column :log_entries, :dm_reward_choice, :integer, default: 0
+  end
+end