+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_09_14_175328) do
+ActiveRecord::Schema.define(version: 2019_09_15_021622) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -119,6 +119,7 @@ ActiveRecord::Schema.define(version: 2019_09_14_175328) do
     t.decimal "tier4_treasure_checkpoints", precision: 6, scale: 1
     t.integer "milestones_gained"
     t.integer "log_format", default: 0
+    t.integer "dm_reward_choice", default: 0
     t.index ["campaign_id"], name: "index_log_entries_on_campaign_id"
     t.index ["user_id"], name: "index_log_entries_on_user_id"
   end